+++ src/models/device.ts
@@ -1,6 +1,5 @@
 import Joi from 'joi';
 
-import { schema as measurementCycleSchema, MeasurementCycle } from './measurement-cycle';
 import { schema as fieldsFromServerSchema, FieldsFromServer } from './fields/fields-from-server';
 
 const schema = Joi.object().keys({
@@ -9,7 +8,7 @@ const schema = Joi.object().keys({
   supplierDeviceIdentifier: Joi.string().required().example('390044000351352237353037').description('Should be unique within the connectivity environment'),
   deviceTypeHashId: Joi.string().required().example('wasd2'),
   fields: fieldsFromServerSchema.required().example({}),
-  measurementCycle: measurementCycleSchema.allow(null).required(),
+  measurementCycle: Joi.valid(null).default(null), // for legacy purposes only
   nextReportBefore: Joi.date().allow(null).required().example('2019-12-31T15:25Z'),
   lastOnlineAt: Joi.date().allow(null).required().example('2019-12-31T15:23Z'),
   validated: Joi.boolean().required().example(true),
@@ -23,7 +22,6 @@ interface Device {
   supplierDeviceIdentifier: string;
   deviceTypeHashId: string;
   fields: FieldsFromServer;
-  measurementCycle: MeasurementCycle | null;
   nextReportBefore: Date | null;
   lastOnlineAt: Date | null;
   validated: boolean;