+++ src/models/supplier-activities/all-activities.ts
@@ -15,7 +15,6 @@ import { schema as parseReportSchema, ParseReport } from './parse-report';
 import { schema as scheduleCommandSchema, ScheduleCommand } from './schedule-command';
 import { schema as sendRequestSchema, SendRequest } from './send-request';
 import { schema as setDeviceFieldsSchema, SetDeviceFields } from './set-device-fields';
-import { schema as setReportingFrequencySchema, SetReportingFrequency } from './set-reporting-frequency';
 import { schema as expectNextReportBeforeSchema, ExpectNextReportBefore } from './expect-next-report-before';
 
 type AllActivities = DropCommand
@@ -33,7 +32,6 @@ type AllActivities = DropCommand
   | ScheduleCommand
   | SendRequest
   | SetDeviceFields
-  | SetReportingFrequency
   | ExpectNextReportBefore;
 
 const schema = (apiVersion: number): Joi.AlternativesSchema => Joi.alternatives().try(
@@ -52,7 +50,6 @@ const schema = (apiVersion: number): Joi.AlternativesSchema => Joi.alternatives(
   scheduleCommandSchema,
   sendRequestSchema,
   setDeviceFieldsSchema,
-  setReportingFrequencySchema,
   expectNextReportBeforeSchema,
 );